# docs/swagger.json
{
    "components": {
        "schemas": {
            "apperrors.FieldError": {
                "properties": {
                    "field": {
                        "type": "string"
                    },
                    "message": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "data": {
                "properties": {
                    "data": {
                        "items": {
                            "$ref": "#/components/schemas/response.Role"
                        },
                        "type": "array"
                    }
                },
                "type": "object"
            },
...
            "request.Login": {
                "properties": {
                    "email": {
                        "type": "string"
                    },
                    "password": {
                        "description": "No min= here. Rejecting a 3-character password at login with \"must be at\nleast 8 characters\" tells an attacker the password was too short to be\nreal, and annoys a legitimate user whose password predates the rule.\nLogin either matches or it does not.",
                        "type": "string"
                    }
                },
                "required": [
                    "email",
                    "password"
                ],
                "type": "object"
            },
            "request.Logout": {
                "properties": {
                    "refreshToken": {
                        "type": "string"
                    }
                },
                "required": [
                    "refreshToken"
                ],
                "type": "object"
            },
            "request.Refresh": {
                "properties": {
                    "refreshToken": {
                        "type": "string"
                    }
                },
                "required": [
                    "refreshToken"
                ],
                "type": "object"
            },
...
            "request.ResendVerification": {
                "properties": {
                    "email": {
                        "type": "string"
                    }
                },
                "required": [
                    "email"
                ],
                "type": "object"
            },
...
            "response.Pagination": {
                "description": "Pagination is present only on list endpoints.",
                "properties": {
                    "hasNext": {
                        "type": "boolean"
                    },
                    "hasPrev": {
                        "type": "boolean"
                    },
                    "page": {
                        "type": "integer"
                    },
                    "perPage": {
                        "type": "integer"
                    },
                    "total": {
                        "type": "integer"
                    },
                    "totalPages": {
                        "type": "integer"
                    }
                },
                "type": "object"
            },
...
            "response.Role": {
                "properties": {
                    "description": {
                        "type": "string"
                    },
                    "id": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "permissions": {
                        "items": {
                            "$ref": "#/components/schemas/response.Permission"
                        },
                        "type": "array",
                        "uniqueItems": false
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "BearerAuth": {
                "description": "Send the access token as: Bearer \u003ctoken\u003e",
                "in": "header",
                "name": "Authorization",
                "type": "apiKey"
            }
        }
    },
    "info": {
        "contact": {
            "name": "the-bughex-code"
        },
        "description": "A production-ready Go backend: JWT authentication, RBAC, and PostgreSQL.\nEvery response uses the same envelope: success, message, data, errors,\npagination, meta, timestamp, requestId.",
        "license": {
            "name": "MIT"
        },
        "title": "Backend API",
        "version": "1.0"
    },
...
    "paths": {
        "/api/v1/auth/login": {
            "post": {
                "description": "Returns an access token (short-lived, stateless) and a refresh token\n(long-lived, revocable). An unknown email and a wrong password return\nthe identical error, in the same amount of time.",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "oneOf": [
                                    {
                                        "type": "object"
                                    },
                                    {
                                        "$ref": "#/components/schemas/request.Login",
                                        "summary": "body",
                                        "description": "Credentials"
                                    }
                                ]
                            }
                        }
                    },
                    "description": "Credentials",
                    "required": true
                },
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/data"
                                        }
                                    ],
                                    "properties": {
                                        "data": {
                                            "description": "Data carries the payload on success, and is null on failure."
                                        },
                                        "errors": {
                                            "description": "Errors carries per-field validation failures, and is null otherwise.\nThis is the only error detail ever echoed back verbatim.",
                                            "items": {
                                                "$ref": "#/components/schemas/apperrors.FieldError"
                                            },
                                            "type": "array",
                                            "uniqueItems": false
                                        },
                                        "message": {
                                            "description": "Message is human-readable and safe to display. It is not stable; do not\nbranch on it.",
                                            "type": "string"
                                        },
                                        "meta": {
                                            "additionalProperties": {},
                                            "description": "Meta is an escape hatch for endpoint-specific extras that do not deserve\na top-level field.",
                                            "type": "object"
                                        },
                                        "pagination": {
                                            "$ref": "#/components/schemas/response.Pagination"
                                        },
                                        "requestId": {
                                            "description": "RequestID ties this response to the server logs. When a user reports a\nbug, this string finds the exact request.",
                                            "type": "string"
                                        },
                                        "success": {
                                            "description": "Success is the single field a client should branch on. It is redundant\nwith the HTTP status code on purpose: some proxies and SDKs swallow the\nstatus, and a body that says success=false is impossible to misread.",
                                            "type": "boolean"
                                        },
                                        "timestamp": {
                                            "description": "Timestamp is the server's UTC clock at the moment of the response. Useful\nfor diagnosing client clock skew, which breaks token expiry checks.",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/response.Envelope"
                                }
                            }
                        },
                        "description": "INVALID_CREDENTIALS"
                    },
                    "403": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/response.Envelope"
                                }
                            }
                        },
                        "description": "ACCOUNT_DISABLED"
                    },
                    "429": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/response.Envelope"
                                }
                            }
                        },
                        "description": "RATE_LIMITED"
                    }
                },
                "summary": "Sign in",
                "tags": [
                    "auth"
                ]
            }
        },
        "/api/v1/auth/logout": {
            "post": {
                "description": "Revokes one refresh token, ending that session. The access token remains\nvalid until it expires; the client must discard it.",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "oneOf": [
                                    {
                                        "type": "object"
                                    },
                                    {
                                        "$ref": "#/components/schemas/request.Logout",
                                        "summary": "body",
                                        "description": "Refresh token"
                                    }
                                ]
                            }
                        }
                    },
                    "description": "Refresh token",
                    "required": true
                },
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/response.Envelope"
                                }
                            }
                        },
                        "description": "OK"
                    }
                },
                "summary": "Sign out",
                "tags": [
                    "auth"
                ]
            }
        },
        "/api/v1/auth/refresh": {
            "post": {
                "description": "Consumes the refresh token and issues a NEW one (rotation). Presenting an\nalready-consumed token revokes every session for that user, because it\nmeans the token was copied.\nThis endpoint is intentionally unauthenticated: the access token has expired.",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "oneOf": [
                                    {
                                        "type": "object"
                                    },
                                    {
                                        "$ref": "#/components/schemas/request.Refresh",
                                        "summary": "body",
                                        "description": "Refresh token"
                                    }
                                ]
                            }
                        }
                    },
                    "description": "Refresh token",
                    "required": true
                },
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/data"
                                        }
                                    ],
                                    "properties": {
                                        "data": {
                                            "description": "Data carries the payload on success, and is null on failure."
                                        },
                                        "errors": {
                                            "description": "Errors carries per-field validation failures, and is null otherwise.\nThis is the only error detail ever echoed back verbatim.",
                                            "items": {
                                                "$ref": "#/components/schemas/apperrors.FieldError"
                                            },
                                            "type": "array",
                                            "uniqueItems": false
                                        },
                                        "message": {
                                            "description": "Message is human-readable and safe to display. It is not stable; do not\nbranch on it.",
                                            "type": "string"
                                        },
                                        "meta": {
                                            "additionalProperties": {},
                                            "description": "Meta is an escape hatch for endpoint-specific extras that do not deserve\na top-level field.",
                                            "type": "object"
                                        },
                                        "pagination": {
                                            "$ref": "#/components/schemas/response.Pagination"
                                        },
                                        "requestId": {
                                            "description": "RequestID ties this response to the server logs. When a user reports a\nbug, this string finds the exact request.",
                                            "type": "string"
                                        },
                                        "success": {
                                            "description": "Success is the single field a client should branch on. It is redundant\nwith the HTTP status code on purpose: some proxies and SDKs swallow the\nstatus, and a body that says success=false is impossible to misread.",
                                            "type": "boolean"
                                        },
                                        "timestamp": {
                                            "description": "Timestamp is the server's UTC clock at the moment of the response. Useful\nfor diagnosing client clock skew, which breaks token expiry checks.",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/response.Envelope"
                                }
                            }
                        },
                        "description": "REFRESH_TOKEN_INVALID, REFRESH_TOKEN_EXPIRED or REFRESH_TOKEN_REVOKED"
                    }
                },
                "summary": "Refresh the access token",
                "tags": [
                    "auth"
                ]
            }
        },
        "/api/v1/auth/register": {
            "post": {
                "description": "Creates a user, grants the default `user` role, and returns the account.\nThe email is trimmed and lowercased before validation.",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "oneOf": [
                                    {
                                        "type": "object"
                                    },
                                    {
                                        "$ref": "#/components/schemas/request.Register",
                                        "summary": "body",
                                        "description": "Registration payload"
                                    }
                                ]
                            }
                        }
                    },
                    "description": "Registration payload",
                    "required": true
                },
                "responses": {
                    "201": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/data"
                                        }
                                    ],
                                    "properties": {
                                        "data": {
                                            "description": "Data carries the payload on success, and is null on failure."
                                        },
                                        "errors": {
                                            "description": "Errors carries per-field validation failures, and is null otherwise.\nThis is the only error detail ever echoed back verbatim.",
                                            "items": {
                                                "$ref": "#/components/schemas/apperrors.FieldError"
                                            },
                                            "type": "array",
                                            "uniqueItems": false
                                        },
                                        "message": {
                                            "description": "Message is human-readable and safe to display. It is not stable; do not\nbranch on it.",
                                            "type": "string"
                                        },
                                        "meta": {
                                            "additionalProperties": {},
                                            "description": "Meta is an escape hatch for endpoint-specific extras that do not deserve\na top-level field.",
                                            "type": "object"
                                        },
                                        "pagination": {
                                            "$ref": "#/components/schemas/response.Pagination"
                                        },
                                        "requestId": {
                                            "description": "RequestID ties this response to the server logs. When a user reports a\nbug, this string finds the exact request.",
                                            "type": "string"
                                        },
                                        "success": {
                                            "description": "Success is the single field a client should branch on. It is redundant\nwith the HTTP status code on purpose: some proxies and SDKs swallow the\nstatus, and a body that says success=false is impossible to misread.",
                                            "type": "boolean"
                                        },
                                        "timestamp": {
                                            "description": "Timestamp is the server's UTC clock at the moment of the response. Useful\nfor diagnosing client clock skew, which breaks token expiry checks.",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        },
                        "description": "Created"
                    },
                    "409": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/response.Envelope"
                                }
                            }
                        },
                        "description": "EMAIL_TAKEN"
                    },
                    "422": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/response.Envelope"
                                }
                            }
                        },
                        "description": "VALIDATION_FAILED"
                    },
                    "429": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/response.Envelope"
                                }
                            }
                        },
                        "description": "RATE_LIMITED"
                    }
                },
                "summary": "Register a new account",
                "tags": [
                    "auth"
                ]
            }
        },
        "/api/v1/auth/resend-verification": {
            "post": {
                "description": "Always returns 200, whether or not the address has an account and\nwhether or not it is already verified. Reporting the difference would\nlet an attacker enumerate your users.\nHeavily rate limited, so it cannot be used as a mail cannon.",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "oneOf": [
                                    {
                                        "type": "object"
                                    },
                                    {
                                        "$ref": "#/components/schemas/request.ResendVerification",
                                        "summary": "body",
                                        "description": "Email address"
                                    }
                                ]
                            }
                        }
                    },
                    "description": "Email address",
                    "required": true
                },
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/response.Envelope"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "422": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/response.Envelope"
                                }
                            }
                        },
                        "description": "VALIDATION_FAILED"
                    },
                    "429": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/response.Envelope"
                                }
                            }
                        },
                        "description": "RATE_LIMITED"
                    }
                },
                "summary": "Resend the verification email",
                "tags": [
                    "auth"
                ]
            }
        },
...
    },
    "openapi": "3.1.0",
    "servers": [
        {
            "url": "localhost:8080/"
        }
    ]
}